publisher: Change to make multiple requests
