--- mqtt-client/src/mqttpublisher.rs
@@ -1,21 +1,0 @@
-use std::error::Error;
-use rumqttc::{self, QoS, AsyncClient};
-
-pub async fn mqtt_publisher(data: String, topic: &str, client: AsyncClient ) -> Result<(),Box<dyn Error + Sync + Send>> {
-    println!("Publishing Fact");
-
-    // Convert the data to bytes to pass as payload for MQTT 
-    let payload = data.as_bytes();
-
-    // Publish the MQTT
-    if let Err(err) = client
-    .publish(topic, QoS::AtLeastOnce, false, payload).await {
-        println!("Failed to publish the MQTT message: {}", err);
-        return Err(Box::new(err));
-    }
-    
-    Ok(())
-}
-
-
-
